Consolidate top level functionality into model and dataset classes
