+++ app.py
@@ -1,13 +1,15 @@
 import streamlit as st
 
 import iml_playground as iml
-from iml_playground import performance, predictions, utils
+from iml_playground import utils
 
-ALT_TITLE_CONFIG = {"fontSize": 14, "offset": 10, "orient": "top", "anchor": "middle"}
-ALT_SCHEME = "tableau10"
+ALTAIR_CONFIG = {
+    "title_config": {"fontSize": 14, "offset": 10, "orient": "top", "anchor": "middle"},
+    "scheme": "tableau10",
+}
 DATASET_TARGET_MAPPER = {
     "car-insurance-cold-calls": "CarInsurance",
-    # "stroke-prediction": "stroke",
+    # "new-dataset-name": "new-dataset-target",
 }
 
 
@@ -20,19 +22,21 @@ def main():
 
     left, right = st.beta_columns(2)
     with left:
-        dataset = st.selectbox(
+        dataset_name = st.selectbox(
             label="Select a dataset",
             options=list(DATASET_TARGET_MAPPER.keys()),
             format_func=lambda s: s.replace("-", " ").title(),
         )
         st.markdown(utils.read_md("dataset.md"))
     with right:
-        train, test = utils.read_train_test(dataset)
-        st.dataframe(test.head(100), height=300)
+        dataset = iml.Dataset(
+            name=dataset_name, target=DATASET_TARGET_MAPPER[dataset_name]
+        )
+        st.dataframe(dataset.sample, height=300)
 
     st.markdown("## Model Predictions and Performance")
 
-    model = iml.Model(train, test, target=DATASET_TARGET_MAPPER[dataset])
+    model = iml.Model(dataset)
 
     left, right = st.beta_columns(2)
     with left:
@@ -43,12 +47,8 @@ def main():
             1.0,
             0.5,
         )
-        chart = predictions.plot_prediction_histogram(
-            model.y_pred,
-            p_min=threshold,
-            p_max=1,
-            title_config=ALT_TITLE_CONFIG,
-            scheme=ALT_SCHEME,
+        chart = model.plot_prediction_histogram(
+            p_min=threshold, p_max=1, altair_config=ALTAIR_CONFIG
         )
         distribution_plot.altair_chart(chart, use_container_width=True)
     with right:
@@ -56,12 +56,9 @@ def main():
 
     left, right = st.beta_columns(2)
     with left:
-        chart = performance.plot_class_performance(
-            y_test=model.y_test,
-            test_preds=model.y_pred,
+        chart = model.plot_class_performance(
             threshold=threshold,
-            title_config=ALT_TITLE_CONFIG,
-            scheme=ALT_SCHEME,
+            altair_config=ALTAIR_CONFIG,
         )
         st.altair_chart(chart, use_container_width=True)
     with right:
@@ -74,7 +71,7 @@ def main():
         st.markdown(utils.read_md("feature_importance.md"))
     with right:
         imp = iml.FeatureImportance(model=model, top_n=10)
-        chart = imp.plot(title_config=ALT_TITLE_CONFIG)
+        chart = imp.plot(altair_config=ALTAIR_CONFIG)
         st.altair_chart(chart, use_container_width=True)
 
     st.markdown("## Global Effects")
@@ -98,7 +95,7 @@ def main():
             model=model,
             method=global_effects_method,
             feature=global_effects_feature,
-        ).plot(title_config=ALT_TITLE_CONFIG)
+        ).plot(altair_config=ALTAIR_CONFIG)
         st.altair_chart(chart, use_container_width=True)
