+++ iml_playground/__init__.py
@@ -1,5 +1,6 @@
+from .dataset import Dataset
 from .global_effects import GlobalEffects
 from .importance import FeatureImportance
 from .model import Model
 
-__all__ = ["FeatureImportance", "GlobalEffects", "Model"]
+__all__ = ["Dataset", "FeatureImportance", "GlobalEffects", "Model"]
